fix(@h6s/table): Rename tabler to table
