+++ packages/table/src/core/renderer/getHeaderCount.ts
@@ -1,5 +1,5 @@
 import { getChildrenCount } from '../../core/renderer/getChildrenCount'
-import { RendererModel } from '../../types/tabler'
+import { RendererModel } from '../../types/table'
 
 export function getHeaderCount<RowData>(model: RendererModel<RowData>): number {
   return model.reduce((acc, result) => {
