+++ packages/table/src/index.ts
@@ -1,4 +1,4 @@
-export * from './core/TablerCore'
+export * from './core/TableCore'
 export { composeDataset } from './helpers/composeDataset'
 export * from './react'
 export * from './types'
